Use Long prices & add cart DTOs
Change price type from double to Long in Game entity and GameDTO to support Stripe unit_amount as integer cents. (Also changed Price in data base from double to Big Int) Add CartDTO and CartItemDTO to accept cart payloads. Update CheckoutController to accept a CartDTO, build Stripe SessionCreateParams dynamically for each cart item (fetching games via gameService and using their Long price), and restrict CORS to http://localhost:5173. Also add Optional handling for game lookup and adjust imports. Minor tweaks to GameController: add @CrossOrigin and remove trailing slash from @GetMapping().

+++ src/main/java/com/nscc/onlinestore2/dto/GameDTO.java
@@ -19,7 +19,7 @@ public class GameDTO {
     private String Category;
 
     @NotNull
-    private double Price;
+    private Long Price;
 
     @NotNull
     @NotBlank(message = "Image is Required")
